Move actlike code to maker
